Trying Gatsby Script tag again without strategy to see if client and srr hydration syncs
